pyproject.toml
[tool.poetry]
name = "luigi-demo"
version = "0.1.0"
description = ""
authors = ["Massimo Biagioli <biagiolimassimo@gmail.com>"]
readme = "README.md"
packages = [{include = "luigi_demo"}]

[tool.poetry.dependencies]
python = "^3.9"
wheel = "^0.37.1"
luigi = "^3.1.1"
beautifulsoup4 = "^4.11.1"
requests = "^2.28.1"
pytest = "^7.2.0"
pytest-mock = "^3.10.0"
psycopg2-binary = "^2.9.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
